CI: enviando a criação de chave do Crypto para a pasta server/src/config/ e deixando o rotacionador mais independente #4
