+++ server/src/security/crypto-engine/KeyManager.ts
@@ -1,11 +1,11 @@
-import { type JWK } from "jose";
+// import { type JWK } from "jose";
 
-class KeyManager {
-  protected static keys: { public: JWK; private: JWK; old: JWK } = {
-    public: {} as JWK,
-    private: {} as JWK,
-    old: {} as JWK,
-  };
-}
+// class KeyManager {
+//   protected static keys: { public: JWK; private: JWK; old: JWK } = {
+//     public: {} as JWK,
+//     private: {} as JWK,
+//     old: {} as JWK,
+//   };
+// }
 
-export default KeyManager;
+// export default KeyManager;
